# docs/swagger.yaml
definitions:
...
  event.CreateArtistRequest:
    properties:
      bio:
        type: string
      name:
        maxLength: 150
        type: string
      photo_url:
        type: string
    required:
    - name
    type: object
  event.CreateCategoryRequest:
    properties:
      icon_url:
        type: string
      name:
        maxLength: 100
        type: string
      slug:
        maxLength: 120
        type: string
    required:
    - name
    - slug
    type: object
  event.CreateEventRequest:
    properties:
      banner_url:
        type: string
      category_ids:
        items:
          type: string
        type: array
      description:
        type: string
      organizer_id:
        type: string
      poster_url:
        type: string
      refund_policy:
        type: string
      slug:
        maxLength: 280
        type: string
      terms_and_conditions:
        type: string
      title:
        maxLength: 250
        type: string
      venue_id:
        type: string
    required:
    - organizer_id
    - slug
    - title
    - venue_id
    type: object
  event.CreateSessionRequest:
    properties:
      end_at:
        type: string
      gate_close_at:
        type: string
      gate_open_at:
        type: string
      session_name:
        type: string
      start_at:
        type: string
      status:
        type: string
    required:
    - end_at
    - gate_close_at
    - gate_open_at
    - start_at
    type: object
  event.UpdateCategoryRequest:
    properties:
      icon_url:
        type: string
      name:
        maxLength: 100
        type: string
      slug:
        maxLength: 120
        type: string
    type: object
  event.UpdateEventRequest:
    properties:
      banner_url:
        type: string
      category_ids:
        items:
          type: string
        type: array
      description:
        type: string
      is_featured:
        type: boolean
      organizer_id:
        type: string
      poster_url:
        type: string
      refund_policy:
        type: string
      rejection_reason:
        type: string
      slug:
        maxLength: 280
        type: string
      status:
        type: string
      terms_and_conditions:
        type: string
      title:
        maxLength: 250
        type: string
      venue_id:
        type: string
    type: object
  helpers.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  models.Artist:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      photo_url:
        type: string
    type: object
  models.Event:
    properties:
      approved_at:
        type: string
      approved_by:
        type: string
      banner_url:
        type: string
      category_ids:
        items:
          type: string
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_featured:
        type: boolean
      max_price:
        type: number
      min_price:
        type: number
      organizer_id:
        type: string
      poster_url:
        type: string
      refund_policy:
        type: string
      rejection_reason:
        type: string
      slug:
        type: string
      status:
        type: string
      terms_and_conditions:
        type: string
      title:
        type: string
      updated_at:
        type: string
      venue_id:
        type: string
    type: object
  models.EventCategory:
    properties:
      icon_url:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  models.EventSession:
    properties:
      created_at:
        type: string
      end_at:
        type: string
      event_id:
        type: string
      gate_close_at:
        type: string
      gate_open_at:
        type: string
      id:
        type: string
      session_name:
        type: string
      start_at:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Konsera ticketing platform
  title: Konsera API
  version: "1.0"
paths:
  /artists:
    post:
      consumes:
      - application/json
      parameters:
      - description: Artist
        in: body
        name: artist
        required: true
        schema:
          $ref: '#/definitions/event.CreateArtistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Artist'
      security:
      - BearerAuth: []
      summary: Create artist
      tags:
      - Artists
...
  /event-categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EventCategory'
            type: array
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Event Categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/event.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EventCategory'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Event Categories
  /event-categories/{category_id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - Event Categories
    get:
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EventCategory'
      security:
      - BearerAuth: []
      summary: Get category
      tags:
      - Event Categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      - description: Category changes
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/event.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EventCategory'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Event Categories
  /events:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
      security:
      - BearerAuth: []
      summary: List events
      tags:
      - Events
    post:
      consumes:
      - application/json
      parameters:
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/event.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Event'
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - Events
  /events/{event_id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - Events
    get:
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
      security:
      - BearerAuth: []
      summary: Get event
      tags:
      - Events
    put:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Event changes
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/event.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - Events
  /events/{event_id}/sessions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Session
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/event.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EventSession'
      security:
      - BearerAuth: []
      summary: Create event session
      tags:
      - Event Sessions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"